# frontend/generated/openapi-current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/accounts": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "List Accounts",
        "description": "Get all active accounts via AccountService.",
        "operationId": "list_accounts_api_v1_accounts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccountDetailDTO"
                  },
                  "type": "array",
                  "title": "Response List Accounts Api V1 Accounts Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Create Account",
        "description": "Create a new account via AccountService.",
        "operationId": "create_account_api_v1_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Account Api V1 Accounts Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{account_id}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get Account",
        "description": "Get account details via AccountService.",
        "operationId": "get_account_api_v1_accounts__account_id__get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDetailDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "accounts"
        ],
        "summary": "Update Account",
        "description": "Update account via AccountService.",
        "operationId": "update_account_api_v1_accounts__account_id__put",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Account Api V1 Accounts  Account Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "accounts"
        ],
        "summary": "Deactivate Account",
        "description": "Soft delete account via AccountService.",
        "operationId": "deactivate_account_api_v1_accounts__account_id__delete",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Deactivate Account Api V1 Accounts  Account Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/accounts/manage": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Api Get Managed Accounts",
        "description": "Get all persistently stored accounts.",
        "operationId": "api_get_managed_accounts_api_accounts_manage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Api Get Managed Accounts Api Accounts Manage Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Api Create Managed Account",
        "description": "Create a new persistent account.",
        "operationId": "api_create_managed_account_api_accounts_manage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Api Create Managed Account Api Accounts Manage Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/manage/{account_id}": {
      "put": {
        "tags": [
          "accounts"
        ],
        "summary": "Api Update Managed Account",
        "description": "Update an existing account.",
        "operationId": "api_update_managed_account_api_accounts_manage__account_id__put",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api Update Managed Account Api Accounts Manage  Account Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "accounts"
        ],
        "summary": "Api Delete Managed Account",
        "description": "Soft delete an account.",
        "operationId": "api_delete_managed_account_api_accounts_manage__account_id__delete",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api Delete Managed Account Api Accounts Manage  Account Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{account_id}/balance": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Api Get Account Balance",
        "description": "Get computed balance for an account.",
        "operationId": "api_get_account_balance_api_accounts__account_id__balance_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Api Get Account Balance Api Accounts  Account Id  Balance Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/accounts/{account_id}/running-balance": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Api Get Account Running Balance",
        "description": "Get running balance for an account.",
        "operationId": "api_get_account_running_balance_api_accounts__account_id__running_balance_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Api Get Account Running Balance Api Accounts  Account Id  Running Balance Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "bank": {
            "type": "string",
            "title": "Bank"
          },
          "account_type": {
            "type": "string",
            "title": "Account Type",
            "default": "savings"
          },
          "balance_paise": {
            "type": "integer",
            "title": "Balance Paise"
          },
          "account_number_last4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Number Last4"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "bank",
          "balance_paise"
        ],
        "title": "AccountCreate",
        "description": "Account creation request."
      },
      "AccountCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "bank": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Bank"
          },
          "account_type": {
            "type": "string",
            "maxLength": 20,
            "title": "Account Type",
            "default": "savings"
          },
          "balance_paise": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Balance Paise",
            "description": "Initial balance in paise",
            "default": 0
          },
          "account_number_last4": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Number Last4",
            "description": "Last 4 digits of account number"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "bank"
        ],
        "title": "AccountCreateRequest",
        "description": "Account creation request."
      },
      "AccountDetailDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Account identifier"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Account name"
          },
          "type": {
            "type": "string",
            "enum": [
              "savings",
              "current",
              "credit_card",
              "investment",
              "loan",
              "other"
            ],
            "title": "Type",
            "description": "Account type"
          },
          "institution": {
            "type": "string",
            "title": "Institution",
            "description": "Bank or institution name"
          },
          "balance_paise": {
            "type": "integer",
            "title": "Balance Paise",
            "description": "Current balance in paise"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "description": "Currency code",
            "default": "INR"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "closed"
            ],
            "title": "Status",
            "description": "Account status"
          },
          "account_number_last4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Number Last4",
            "description": "Last 4 digits"
          },
          "opened_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opened Date",
            "description": "Account opening date (ISO)"
          },
          "closed_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed Date",
            "description": "Account closing date (ISO)"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Account notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "institution",
          "balance_paise",
          "status"
        ],
        "title": "AccountDetailDTO",
        "description": "Detailed account information."
      },
...
      "AccountUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "bank": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank"
          },
          "account_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Type"
          },
          "balance_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance Paise"
          },
          "account_number_last4": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Number Last4"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "AccountUpdate",
        "description": "Account update request."
      },
      "AccountUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "bank": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank"
          },
          "account_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Type"
          },
          "balance_paise": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance Paise"
          },
          "account_number_last4": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Number Last4"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "AccountUpdateRequest",
        "description": "Account update request."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}